refactor to fire sendEvent on checkout start
