+++ support-frontend/assets/pages/digital-subscription-landing/components/paymentSelection/helpers/paymentSelection.tsx
@@ -14,8 +14,8 @@ import {
 } from 'helpers/productPrice/billingPeriods';
 import type {
 	BillingPeriod,
-	DigitalBillingPeriod,
-	DigitalGiftBillingPeriod,
+	// DigitalBillingPeriod,
+	// DigitalGiftBillingPeriod,
 } from 'helpers/productPrice/billingPeriods';
 import {
 	getAdverbialSubscriptionDescription,
@@ -30,7 +30,6 @@ import type {
 	ProductPrice,
 	ProductPrices,
 } from 'helpers/productPrice/productPrices';
-import type { Promotion } from 'helpers/productPrice/promotions';
 import { getAppliedPromo } from 'helpers/productPrice/promotions';
 import type { SubscriptionProduct } from 'helpers/productPrice/subscriptions';
 import {
@@ -40,10 +39,6 @@ import {
 } from 'helpers/productPrice/subscriptions';
 import { gaEvent } from 'helpers/tracking/googleTagManager';
 import type { OphanComponentType } from 'helpers/tracking/ophan';
-import {
-	sendEventIds,
-	sendEventWithCurrency,
-} from 'helpers/tracking/quantumMetric';
 import type { Option } from 'helpers/types/option';
 import { getDigitalCheckout } from 'helpers/urls/externalLinks';
 
@@ -82,38 +77,6 @@ export const getProductPrice = (
 	}
 
 	throw new Error('getProductPrice: product price unavailable');
-};
-
-export const getAnnualValue = (
-	fullPrice: number,
-	promotion: Promotion | undefined,
-	billingPeriod: DigitalBillingPeriod | DigitalGiftBillingPeriod,
-): number => {
-	const periodMultiplier: Record<
-		DigitalBillingPeriod | DigitalGiftBillingPeriod,
-		number
-	> = {
-		Annual: 1,
-		Monthly: 12,
-		Quarterly: 4,
-	};
-	const fullPriceInPenceCents = fullPrice * 100;
-	const fullAnnualPrice =
-		fullPriceInPenceCents * periodMultiplier[billingPeriod];
-
-	if (!promotion?.discountedPrice || !promotion.numberOfDiscountedPeriods) {
-		return fullAnnualPrice;
-	}
-
-	const discountedPriceInPenceCents = promotion.discountedPrice * 100;
-	const discountInPenceCents =
-		fullPriceInPenceCents - discountedPriceInPenceCents;
-
-	const discountedAnnualPrice =
-		fullAnnualPrice -
-		discountInPenceCents * promotion.numberOfDiscountedPeriods;
-
-	return discountedAnnualPrice;
 };
 
 const BILLING_PERIOD = {
@@ -330,19 +293,6 @@ const getPaymentOptions = ({
 			});
 
 			sendTrackingEventsOnClick(trackingProperties)();
-
-			const annualValue = getAnnualValue(
-				fullPrice,
-				promotion,
-				billingPeriod as DigitalBillingPeriod | DigitalGiftBillingPeriod,
-			);
-
-			sendEventWithCurrency(
-				sendEventIds.digiSubSelected,
-				false,
-				currencyId,
-				annualValue,
-			);
 		};
 
 		return orderIsAGift
